requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-decouple==3.4
pytube==10.7.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
wincertstore==0.2
